+++ src/algorithms/myers.rs
@@ -11,13 +11,8 @@
 //!
 //! # Heuristics
 //!
-//! At present this implementation of Myers' does not implement any more advanced
-//! heuristics that would solve some pathological cases.  For instance passing two
-//! large and completely distinct sequences to the algorithm will make it spin
-//! without making reasonable progress.  Currently the only protection in the
-//! library against this is to pass a deadline to the diffing algorithm.
-//!
-//! For potential improvements here see [similar#15](https://github.com/mitsuhiko/similar/issues/15).
+//! This module contains the core Myers implementation. Any higher-level
+//! pre-flight heuristics are implemented outside of this module.
 
 use std::ops::{Index, IndexMut, Range};